+++ src/Compute/Compute/ChangeLog.md
@@ -20,6 +20,7 @@
 
 -->
 ## Upcoming Release
+* New cmdlet `Start-AzVmssRollingExtensionUpgrade`.
 
 ## Version 4.6.0
 * Added `-VmssId` parameter to `New-AzVm`
@@ -27,7 +28,6 @@
 * New cmdlet `Get-AzDiskEncryptionSetAssociatedResource`
 * Added `Tier` and `LogicalSectorSize` optional parameters to the New-AzDiskConfig cmdlet. 
 * Added `Tier`, `MaxSharesCount`, `DiskIOPSReadOnly`, and `DiskMBpsReadOnly` optional parameters to the `New-AzDiskUpdateConfig` cmdlet. 
-* New cmdlet `Start-AzVmssRollingExtensionUpgrade`.
 
 ## Version 4.5.0
 * Fixed issue in `Update-ASRRecoveryPlan` by populating FailoverTypes